+++ pkgs/watcher/test/directory_watcher/polling_test.dart
@@ -32,6 +32,39 @@ void main() {
       writeFile('b.txt', contents: 'after');
       await expectModifyEvent('b.txt');
     });
+
+    // A poll does an async directory list then checks mtime on each file. Check
+    // handling of a file that is deleted between the two.
+    test('deletes during poll', () async {
+      await startWatcher();
+
+      for (var i = 0; i != 300; ++i) {
+        writeFile('$i');
+      }
+      // A series of deletes with delays in between for 300ms, which will
+      // intersect with the 100ms polling multiple times.
+      for (var i = 0; i != 300; ++i) {
+        deleteFile('$i');
+        await Future<void>.delayed(const Duration(milliseconds: 1));
+      }
+
+      final events =
+          await takeEvents(duration: const Duration(milliseconds: 500));
+
+      // Events should be adds and removes that pair up, with no modify events.
+      final adds = <String>{};
+      final removes = <String>{};
+      for (var event in events) {
+        if (event.type == ChangeType.ADD) {
+          adds.add(event.path);
+        } else if (event.type == ChangeType.REMOVE) {
+          removes.add(event.path);
+        } else {
+          fail('Unexpected event: $event');
+        }
+      }
+      expect(adds, removes);
+    });
   });
 
   // Also test with delayed writes and real mtimes.